feat: add authorization to core / server-side components
core:
  - add authorization logic to usecases
  - make authorization enforcement optional
  - remove possibility to update an entrypoint urlMatcher and appId, as it
    might cause an escalation of priviliges
  - remove possbility to delete an app if it has linked entrypoints
  - make sure the storages updateOne methods are not called with objects
    containing more properties than expected

others:
  - update accordingly to the core module

+++ core/src/usecases/DeleteApp.ts
@@ -1,32 +1,35 @@
-import { AppNotFoundError } from "../common/errors";
+import { isEmpty } from "lodash";
+
+import { AppHasEntrypointsError, AppNotFoundError } from "../common/errors";
 import Usecase from "../common/Usecase";
 import { Operation } from "../entities/OperationLog";
 
 export default class DeleteApp extends Usecase {
     async exec(id: string): Promise<void> {
-        // Ensure the request is authenticated
-        this.authorizer.ensureAuthenticated();
+        // Auth check
+        this.authorizer.ensureCanDeleteApp(id);
 
         const toBeDeletedApp = await this.storages.apps.findOne(id);
-        const toBeDeletedEntrypoints = await this.storages.entrypoints.findManyByAppId(
-            id
-        );
 
         // Ensure the app exists
         if (!toBeDeletedApp) {
             throw new AppNotFoundError(id, "id");
         }
 
-        // Delete linked entrypoints
-        await this.storages.entrypoints.deleteManyByAppId(id);
+        // Ensure the app has no linked entrypoints
+        const linkedEntrypoints = await this.storages.entrypoints.findManyByAppId(
+            id
+        );
+        if (!isEmpty(linkedEntrypoints)) {
+            throw new AppHasEntrypointsError(id);
+        }
 
         // Delete the app
         await this.storages.apps.deleteOne(id);
 
         // Log the operation
         await this.operationLogger.logOperation(Operation.deleteApp, {
-            deletedApp: toBeDeletedApp,
-            deletedEntrypoints: toBeDeletedEntrypoints
+            deletedApp: toBeDeletedApp
         });
     }
 }
